COMP: Fix narrowing warnings in vtkMRMLTransformStorageNode
This commit fixes the following warnings:

/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx: In instantiation of ‘bool SetVTKBSplineFromITK(vtkObject*, vtkOrientedBSplineTransform*, itk::TransformBaseTemplate<double>::Pointer, itk::TransformBaseTemplate<double>::Pointer) [with T = double; itk::TransformBaseTemplate<double>::Pointer = itk::SmartPointer<itk::TransformBaseTemplate<double> >]’:
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:611:88:   required from here
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<double>::<anonymous>.itk::Array<double>::<anonymous>.vnl_vector<T>::operator[]<double>(0u)’ from ‘const double’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]
   const int gridSize[3]={transformFixedParamsItk[0], transformFixedParamsItk[1], transformFixedParamsItk[2]};
                                                                                                            ^
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<double>::<anonymous>.itk::Array<double>::<anonymous>.vnl_vector<T>::operator[]<double>(1u)’ from ‘const double’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<double>::<anonymous>.itk::Array<double>::<anonymous>.vnl_vector<T>::operator[]<double>(2u)’ from ‘const double’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:396:12: warning: unused variable ‘linear’ [-Wunused-variable]
     double linear[VTKDimension][VTKDimension] = { {1,0,0}, {0,1,0}, {0,0,1}};
            ^
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:397:12: warning: unused variable ‘offset’ [-Wunused-variable]
     double offset[VTKDimension] = {0,0,0};
            ^
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx: In instantiation of ‘bool SetVTKBSplineFromITK(vtkObject*, vtkOrientedBSplineTransform*, itk::TransformBaseTemplate<double>::Pointer, itk::TransformBaseTemplate<double>::Pointer) [with T = float; itk::TransformBaseTemplate<double>::Pointer = itk::SmartPointer<itk::TransformBaseTemplate<double> >]’:
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:612:88:   required from here
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<float>::<anonymous>.itk::Array<float>::<anonymous>.vnl_vector<T>::operator[]<float>(0u)’ from ‘const float’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]
   const int gridSize[3]={transformFixedParamsItk[0], transformFixedParamsItk[1], transformFixedParamsItk[2]};
                                                                                                            ^
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<float>::<anonymous>.itk::Array<float>::<anonymous>.vnl_vector<T>::operator[]<float>(1u)’ from ‘const float’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]
/path/to/Slicer/Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx:316:108: warning: narrowing conversion of ‘(& transformFixedParamsItk)->itk::OptimizerParameters<float>::<anonymous>.itk::Array<float>::<anonymous>.vnl_vector<T>::operator[]<float>(2u)’ from ‘const float’ to ‘const int’ inside { } is ill-formed in C++11 [-Wnarrowing]

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=23155
git-svn-id: http://svn.slicer.org/Slicer4/trunk@23155 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Core/vtkMRMLTransformStorageNode.cxx
@@ -313,7 +313,7 @@ template <typename T> bool SetVTKBSplineFromITK(vtkObject* self,
     }
 
   // Get grid parameters from the fixed parameters
-  const int gridSize[3]={transformFixedParamsItk[0], transformFixedParamsItk[1], transformFixedParamsItk[2]};
+  const double gridSize[3]={transformFixedParamsItk[0], transformFixedParamsItk[1], transformFixedParamsItk[2]};
   double gridSpacing[3]={transformFixedParamsItk[6], transformFixedParamsItk[7], transformFixedParamsItk[8]};
   double gridOrigin_LPS[3]={transformFixedParamsItk[3], transformFixedParamsItk[4], transformFixedParamsItk[5]};
   vtkNew<vtkMatrix4x4> gridDirectionMatrix_LPS;
